+++ kiari/cli/_constants/common_option_groups.py
@@ -88,7 +88,7 @@ COMMON_OPTION_GROUPS: list[OptionGroupDict] = [
             "--agent-id",
             "--runner-id",
             "--language",
-            "--time-zone",
+            "--timezone",
             "--currency",
         ],
     },